Remove namespace, utcseconds, add msg header, rename purge -> flush
1) Neither namespace or utcseconds are *really* needed, and removing
them simplifies the scripts and increases performance substantially.
If time is needed in the client, it can be added to the message in
whatever format required. For now we should be trying to do the least
possible and build additional features on top of the foundation.

2) Add a header to messages to allow channels and ids to be easily
parsed in client libraries. This is important when consuming using
globbing since the actual channel published on can't be inferred.

3) "Flush" makes more sense than "purge". Otherwise unchanged.